requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
cipher==0.1
colorama==0.4.6
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
protobuf==4.24.4
pytube2==15.0.10
requests==2.32.3
setuptools==75.1.0
urllib3==2.2.3
wheel~=0.44.0
